Specs: Automated VLAN device creation
If an instance gets attached to an VLAN network, the instance
user needs to configure that VLAN manually from within its
instances operating system using a VLAN device.

This blueprint proposes to take use of the cloud-init service
in combination with the ConfigDrive to do this configuration
automatically.

Change-Id: Ifab56a67308eb6f39343c675b715f11aa0a5fab1
Implements: blueprint neutron-integration-vlan