Introduce client_proxy to improve testability with fakezhmcclient
The DPM driver utilizes the python-zhmcclient to
interact with the REST API of the HMC of a system z.
To enable unit-testing, a fakezhmcclient got introduced,
which mocks the behavior of the real zhmcclient. The
usage of the mocking was spread over multiple modules
with duplicated code.
This change introduces a client_proxy module which
allows us to mock the usage of the fakezhmcclient at
one central place in unit tests.

Change-Id: Ib7a80e89d6a4a48c21f536dd1f6937d23f26bcf9