Cut the dependency to zhmcclient internals
We imported internals from the zhmcclient, for example, imports like this:

    from zhmcclient._manager import BaseManager
    from zhmcclient._resource import BaseResource

These types are not part of the public interface of the zhmcclient. This
breaks us when the zhmcclient changes its internals in a backwards
incompatible way. For example, commit ac5dbc45b was necessary to fix us.

This change removes the dependency to zhmcclient internals and introduces
as less code as possible to keep the unit tests running.

Closes-Bug: 1656230
Change-Id: I3229d684067a5dbf0e98a31909cf8bca580da908