Install stable zhmcclient via requirements.txt
Till now, zhmcclient is installed from github masterbranch using
pip. As zhcclient is now an official OpenStack requirement [1]
we can switch to install the latest stable version via
requiremts.txt.

To allow newer versions in the future, an update to the
upper-constraints.txt in the OpenStack requirements repo is
required.

To bump the minimum just increase the version in requirements.txt
of nova-dpm and make sure, that newer versions are allowed
by OpenStack requirements (see above).

[1] https://review.openstack.org/413689

Change-Id: I0da5f9271ee3391a4648064207eab5b4fb7e26fb
Closes-Bug: #1655905