Add nic information to partitions boot params
In order to bring up networking in a dpm instance correctly,
additional data needs to be available from inside the instance.
As a config drive is not yet available, the DPM
'boot-os-specific-parameters' property is exploitet.
The set value will be appended to the instance kernels cmdine and be
available under /proc/cmdline

For each NIC the following information will be added to this
parameter:

  <devno>,<port-no>,<mac>;

Where the <port-no> will default to 0 until support has been added to
networking-dpm.

The 'boot-os-specific-parameters' value is limited to 256 chars.
Therefore a checking is introduced that throws an Exception if
the maximum limit of NICs is exceeded. The limit is 12 NICs/ports
per partition.

Implements: blueprint neutron-integration-set-mac
Change-Id: Ie6f9e1784219feffcefa5bda7194768b32b28a7c