Changed name of cpc_uuid config option to cpc_object_id
Reason for the change is that the HMC API book does not mention
anywhere that the format of the ID is a UUID, and the name of the
resource property is "object-id". This change set changes the
name of the DPM config option "cpc_uuid" to "cpc_object_id",
and the devstack variables CPC_UUID to CPC_OBJECT_ID, and
updates any use of that and the unit testcases accordingly.

Change-Id: I6153ac791d4296101dc00f23dfb7fb7c3d2fee2c
Partial-Bug: 1659304
Signed-off-by: Andreas Maier <maiera@de.ibm.com>