Docs: Add Feature Support Matrix
To give customers a glimpse of what to expect from this driver,
this change adds a feature support matrix.
As the nova-dpm driver is out-of-tree, we cannot add it to the
in-tree documentation at
    http://docs.openstack.org/developer/nova/support-matrix.html
Having *all* drivers in this change would add too much maintenance
and manual sync-up between the two repositories (novavs. nova-dpm),
therefore we list only the dpm-driver here.

Closes bug: 1657700

Change-Id: Ie65cb50dc54ce8116f608c3a5280407cf820a2b0