Docs: Add Release Notes
To give customers a glimpse of what to expect from each release,
this change adds a way to add release notes.

The "normal" OpenStack projects have two different sphinx directories:
* doc/source/
* releasenotes/source
That's reasonable for them, as they publish to two different locations.

As an out-of-tree driver, we cannot use the usual release notes
location at http://docs.openstack.org/releasenotes/ All our docs go
to http://nova-dpm.rtfd.io/

This made it necessary to *not* use the default release notes
directory "releasenotes/source/" like the other projects.
Otherwise we would have two different RTD projects, one for the
"normal" docs and another one for the release notes.

As we only have one place for documentation (including release notes)
it didn't make sense to have a separate tox target to create those
release notes. They are included when the docs are built with:

    $ tox -e docs

This slight deviation from the upstream defaults made it necessary
to explain the dealing with reno release notes in a developer reference
(devref).

The one added release note is only to double-check if everything
behaves as expected. It has to be updated before the next release.

Closes bug: 1657842
Change-Id: I0d2d10ea2d91bb5193aa1092997db310b277878b