Replace CONF mocking with CONF.set_override
As we want to get this code in-tree of the Nova repository one day, we
should be in line with their conventions as much as possible. This will
make the transition effort in the future much easier.
One of the conventions Nova has, is the usage of the "set_override"
method offered by "oslo.config" to override specific config options.
This allows to write unit tests which test for a certain behavior under
different configurations.
This change replaces the way we mocked the configuration with the way
Nova does. It does so by introducing a "flags" method in the base
TestCase class to allow every derived TestCase class to simply call
"self.flags()".
The name "flags" is a historical term in Nova and used to describe
config options. Using a different method name would make the transition
to going in-tree harder.

This is also a preparation step which makes resolving bug 1655043
easier.

partial bug: 1655043

Change-Id: I9df123d9283eb10632919f650500c38060b2954a