Remove config option [dpm].host
The DPM driver manages so-called "cpc-subsets". Basically, this is
a logically portion cut out of a system z CPC. For the Nova driver,
this emulates a "host". This results in this multiplicity:

    DPM Driver 1 --- 1 CPC-subset (=host) * --- 1 CPC

Having the two config options "[DEFAULT].host" and "[dpm].host", which
mean the same thing, could be confusing for a user. Also, this can
lead to issues if both are not in sync, for example during the
Nova<->Neutron communication where the NICs get created.

To resolve this, this change removes the config option "[dpm].host"
and uses "[DEFAULT].host" only. As the config option wasn't yet part
of any release, we don't need to do the usual deprecation policy.

closes bug: 1655043

Change-Id: Ibb2c23e248c4390e2ec19e7aa9811abdc24fea25