Use os-dpm
Os-dpm [1] is a shared library between nova-dpm and networking-dpm.
For now it contains the common config options between those 2
projects.

This patch switches over to use os-dpm. Config options that where
duplicates between nova-dpm and os-dpm, are now removed from
nova-dpm. This affects

* hmc
* hmc_password
* hmc_username

The config option *cpc_uuid* was not touched so far. os-dpm
introduces a new config option *cpc_object_id* which should be used
in the future. But this requires a follow up patch. For now both options
exist, whereas nova-dpm is still using the cpc_uui. This affects

* hmc
* hmc_password
* hmc_username

The config option *cpc_uuid* was not touched so far. os-dpm
introduces a new config option *cpc_object_id* which should be used
in the future. But this requires a follow up patch. For now both options
exist, whereas nova-dpm is still using the cpc_uuid.

Change-Id: If0b877101a9603ce04a5e0aaba3d4b9359b8d0fd
Closes-Bug: #1661315