Tests for guest image tools
Adding Testcases for the functions used by guest-image-tools. The
tests are executed by python and integrated into the py27 and py35
test jobs.

Sometimes the guest image tools call external commands
(like ip, zneconf,...). for this case dummy versions of those
executeables are placed in a faked root directory. The test takes
care of exectuing those commands in this directory, therefore the
dummy (bash) executables are chosen over the real ones.
However those dummy executeables cannot replace mocks where a lot
of assertions can be done. As workaround the dummy executeable just
check if it was called with a certain list of arguments. The test
needs to exactly pass those arguments in order to succeed.

Ideally those tests would be functional tests running on a real
system z envrionement. Doing so we could verify if the real
operations are working as well.

Change-Id: I35d0840c2e4bb9d216d57a5927f20b3f0ff182c0