Add Ocata (1.0.0) release notes
We publish our docs on ReadTheDocs (RTD). We also want to use "reno"
as release notes tools. Reno has issues to determine the repo root
when it gets used in an RTD build (see bug 1661319).
The bug is solved in reno>2.0.3 but that will be released *after*
the Ocata release. This means we need an interim solution to have
release notes ready at the same time when the actual release is
happening. This change is this interim solution.

Change-Id: I90c5419b7338b312a469308a59501f3bd0ef02ba
Co-Authored-By: Sreeram Vancheeswaran <sreeram.vancheeswaran@in.ibm.com>