Set exit on error
Since some of the functions are getting called in subshells. Calling
exit in a subshell just quits the subshell but not the whole
programm. As some of the functions should lead to a programm
termination, a fix is required.

Setting "set -e" ensures that the whole programm is being exited in
such cases.

Change-Id: I3269534029e40109e7cec580a83a27f489acbe17