Introduce storage adapter mapping conf object
This patch introduces a oslo config object that can be used by the
physical_storage_adapter_mappings configuration option.

The config object validates the config option along a regex and
returns a tuple (adapter_id, port).

It accepts adapter-ids in upper and lower case. But for users
it returns the adapter-id always in lower case (like desired
by the HMC web services API).

Some parts of this code can be removed once they are available
in a new os-dpm release (e.g. the MultiStringWithKwargsType).

Change-Id: I03319412ad15c608ecc43b0421f0cae42bd6b5dc
Partial-Bug: #1663369