Remove obsolete self._properties var in host.py
This var has not been used anywhere. Maybe it has been implemented
to be prepared for caching. But caching does not make sense in this
context, as host.properties is called regularily to report the
current agent state to the conductor. Therefore the latest state
needs to be captured. Caching would report the old state all the
time.

Change-Id: I9b1dc5cb3f063ffd099bbf10facf52246ec29fef