Use oslo opt for storage_adapter_mappings
Use the oslo config opt object for the
'physical_storage_adapter_mappings' configuration option. This configuration
option returns a list of tuples:

  [(adapter-id, port-id), (adapter-id, port-id)

In addition it validates if the adapter-id is of the right format. If
provided in upper case, it will convert it to lower case (HMC web
services API requires lower case). The config otpion does not check
if an object with this ID exists on the HMC.

Also the port value is verified. Valid values are '0' and '1'.

adapter-id and port can never be 'None'.

Change-Id: Ica7928e1eded0350dcc708eed932e9b7a5401bfc
Partial-Bug: #1663369