Adapting to redesign of async operation methods in zhmcclient
In zhmcclient version 0.11.0 onwards, the return type of all methods on
resource classes that invoke asynchronous operations (i.e. all methods
that have a wait_for_completion parameter), is changed. For
wait_for_completion=True, the JSON object in the 'job-results' field
is now returned, or None if not present (i.e. no result data).
Previously, the complete response was returned as a JSON object [1]

Hence changes as per this patch are needed, so that the nova-dpm code
no longer depends on the return value for when such methods when called
using wait_for_completion=True

[1] https://github.com/zhmcclient/python-zhmcclient/pull/182

Change-Id: Icc73e4bd6de6444cd0843053409074297041d0e5
Closes-bug: #1660905
Signed-off-by: Sreeram Vancheeswaran <sreeram.vancheeswaran@in.ibm.com>