Cleanup spec folder
All the specs are listed in our "Contributing to the project"
section of the docs. Also the templates are shown there. At the
moment we have more templates than specs listed there.

Removing the duplicated template file and replace it by a .dummy file
which is not displayed when building the docs. Now only the common
ocata template is shown.

Change-Id: I7b8af6566b4d422d4e26f433e608706786d7c38d