Fix unit tests with oslo.config "enforce_type" setting
This oslo.config change enforces the type of an option by default:
https://github.com/openstack/oslo.config/commit/3a8ba0f8ea5
But that is not yet released.

This commit in oslo.config should make the transition to default
type enforcement easier, but it also not yet released:
https://github.com/openstack/oslo.config/commit/04a78cc1d0

We use Nova's "flags" method to override config options. That method
did already remove the "enforce_type=True" setting:
https://github.com/openstack/nova/commit/213f7120c

And that's why the Nova change did brake our unit tests.

This change uses the native "set_override" method of oslo.config
which allows us to set "enforce_Type=True" which then does a type
conversion which creates the types we expected.
This change is not the ultimate solution but a short term fix to
unwedge our gate checks.

Change-Id: I0445b22e87aa6434d057264d9b751eb0f2d57f06