Boot from image fails without a clear error message
Booting an instance form image fails somewhere with
a magic error message.But as we do not support the boot
from image case. We should check that at first
when the spawn method is entered or even the
pre_for_spawn method is entered and abort with
an appropriate error message!

Closes-Bug: #1663550

Change-Id: I577b28d034322b109b81c978cb50a299be140510
Signed-off-by: Prabhat Ranjan <pranjank@in.ibm.com>