guest tools: Fix layer2 configuration
The guest image tools do not properly configure a network interface
in the image for layer2. When hipersockets is used, setting the MAC
fails with "Operation not supported". No communication with the
dhcp on the network is possible.

OSA was not affected as layer2 is the default for OSA if nothing is
provided (for hipersockets it's layer3).

Change-Id: I203a8d4e62afb10666e9b131164e77ab4f85d890
Closes-Bug: #1685737