Multi storage adapter support broken
The initator-target map, we're not necessarily picking
the entry for our boot adapter, but just the first one
that is returned by get_partition_wwpns. But in fact we
must not take the first one, but we must take the wwpn
of our selected boot adapter as initiator wwpns
(and then lookup the target wwpns for it).

Closes-Bug: #1662511

Change-Id: I0bf6297042fc821f2c2feb422297fff36a8ef98a
Signed-off-by: Prabhat Ranjan <pranjank@in.ibm.com>