Multi storage adapter support broken
The initator-target map, we're not necessarily picking
the entry for our boot adapter, but just the first one
that is returned by get_partition_wwpns. But in fact
we must not take the first one, but we must take the
wwpn of our selected boot adapter as initiator wwpns
(and then lookup the target wwpns for it).

Closes-Bug: #1662511

Change-Id: Ib5890c9e696e6bfae5d2915f87b4baf151b2ed9a
Signed-off-by: Prabhat Ranjan <pranjank@in.ibm.com>