stable ocata uses master branch of nova
The tox environment downloads nova master even when stable/ocata
branch is checked out in nova-dpm. However in this case also
the stable version of the project should be used!

This patch also contain one more change

What another change?
Ans: In stable/ocata release we did not have "functional"
test enabled. So when we are trying to fix stable/ocata
with launchpad bug-1672438 then jenkins were failing because
we have enabled "functional" test now in pike cycle. So
inorder to not break jenkins for stable/ocata
we have added(enabled) "functional" test.

Closes-Bug: #1672438

Change-Id: Ibcac0eb71f8881b8fa7884e79af7526048192011
Signed-off-by: Prabhat Ranjan <pranjank@in.ibm.com>