Update local.conf
With the current local.conf it is not possible to use floating-ips
to access an instance.

This patch extends and refactors the local.conf, that devstack creates
2 networks. A public network (used for floating ip) - public, and a
private network where the instances get attached to - provider.

In addition the ovs is configured to be able to deal with both networks.
However one manual step is required. The bridge br-provider needs to be
created manually before stack.sh is run.

Change-Id: I776552c41161d53d33ad9c6c750f47ef6d5ccce6