PowerOn of a partition in paused state not working
When we do inband shutdown of partition using
command 'shutdown -P now'
Then partition is not going to 'stopped' state.
It is going to 'PAUSED' state.

In DPM there is no such implementation so that we
can start partition from PAUSED state. So workaround
is we need to stop the partition then start.

This patch also includes the cases when partition is
in "starting/stopping" state.

Closes-Bug: #1678002

Change-Id: I1898b85a8441c3175db0a6de6d620bca19af69ec
Signed-off-by: Prabhat Ranjan <pranjank@in.ibm.com>