Min value of max_memory is been added
512mb is the minimum memory that is required to launch a partition and
rounding off starts only from value greater that 512mb this decides the
min value needs to be set to the config parameter max_memory.
There is a parameter 'reserved_host_memory_mb' in nova.conf which
reserve 512mb of the alloated memory to scale up the partition memory.
This is restricted to '0' to have optimized utilization of the memory.

closes-Bug: 1689267

Change-Id: I493fec40b412feb5303ab7ca9ee2f78c9b9220db