Error handling for ConnectionError
when starting/restarting the nova-dpm service and if it not able to
connect to hmc then through error as soon as possible so
that cloud admin will get to know very early stage.
Means there is no point to retry so many times
when we start/restart service.

But when service is already up and somebody reboot
hmc or doing something with hmc so that service is not able
to connect to hmc then in that case we can increase the retry value.

Closes-Bug: #1665401

Change-Id: I0f09cfcc0a8cb77a450452a7856ffa53a2fb431f
Signed-off-by: Prabhat Ranjan <pranjank@in.ibm.com>