Removal of max_instances config parameter
There is a NumInstancesFilter which can be used to filter the number of
instances that are supposed to be launched on the host this replaces the
work that is planned for max_instances config parameter to perform.
It is removed as the main functionality of this parameter is not yet
implemented in the code.

Change-Id: Ie545a5f74f7b63ea56812577c201fcb443014998