Retrieve PartitionInstance via getter method
In some methods we directly created an instance of the PartitionInstnace object.
The instantiation of this object is now moved into a separate method. This is
even more usefull when we introduce ssc support in the future. In that case
we could simply introduce a SSCPartitionInstance class. The newly created method
would be the only place where the decission between ssc and pure dpm is done.

Change-Id: I06e6d8ee84b7f308f2cdbf120c08e46afd4dd8be