Retrieve boot_hba uri from hba object
In some cases we need the hba object to retrieve the wwpn of
the hba used for boot, in some the element-uri. We had 2 different
ways of gathering this information.

This patch transforms the get_boot_hba_uri method into a more generic
get_boot_hba method. From this object it's easy to retrieve the
required properties 'element-uri' and 'wwpn'.

Change-Id: If518a915408caf5c59d4425b8aa2e0673acbcc19