Change how we set the boot-os-specific-params for a nic
For each nic we need to add some data to the partitions
boot-os-specific-parameters property. Till now we were accumulating
everything to a single string which set the property once. The new
approach is to update the property for each nic.

This leads to more read/write accesses on this partition property. But
on the other hand we're gaining flexibility. E.g. someone might want to
add a NIC after spawn. Or some other data should be added. Those
scenarios were not supported in the past.

This is also helpful to move this code into the partitioninstance
class. In the future this action might not be required for all type of
partitions (e.g. not for ssc).

Change-Id: I0262f2660b3e89b455bb1b1ece016342c8347c80