Remove unused function _build_resources
The usage of this method got removed long time ago with commit [1].
The purpose of this method was to call a cinder function [2] to discover the lun
and mount the block device to the compute node. With DPM this is not required,
as the operating system of the partition is taking care of lun discovery of the
block deivce. The compute node does not see the lun at all.

[1] https://github.com/openstack/nova-dpm/commit/eb2bc803892a17e27e9f7fd1af12f313878fb71a
[2] https://github.com/openstack/nova/blob/a40c00957ef4552681c172913bae1135a1614bbc/nova/compute/manager.py#L1586,L1597

Change-Id: I5c945b79f63d8d6fdccc343de7c72eda43c61476