Use status_timeout when starting a partition
We now specify the status timeout when starting a partition. The
zhmcclient will wait that amount of seconds for the partition
to become active (or another valid start status).

Without explicitly setting it, the zhmcclient will wait for 900 seconds.

In the past we waited that 900 seconds + the 60 extra seconds, as we
were calling partition.wait_for_completion right after the start of
the partition again.

Change-Id: I483efc5808c1a91993dbde936c27b82302640f9e