Modernize doc building process
Adopt the latest and greatest in docs standards, including a separate
requirements file for docs and the use of sphinx-build instead of the
setuptools command.

A call to oslo-config-generator is removed since this is already enabled
by way of the 'oslo_config.sphinxconfiggen' extension. Similarly, a
number of unused settings are removed from 'conf.py'

Change-Id: Ib6fa0276e0a3bffd9b1fbe759b718ec8824490d0
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>